Refactor changes to match basic auth style
