+++ sandbox/libs/authn/src/main/java/org/opensearch/authn/internal/InternalAccessTokenManager.java
@@ -22,6 +22,11 @@ public class InternalAccessTokenManager implements AccessTokenManager {
     }
 
     @Override
+    public void expireToken(final AccessToken token){
+
+    }
+
+    @Override
     public AccessToken generate() {
         return new AccessToken();
     }
